adjust discharge dates according to death date
If the discharge occurs 1 - 5 days after death use the death date in place of the discharge date.
Need to adjust all sc_date_2 (the final discharge date) and only the keydate_2 when this is the last record in the full episode.
Use the `death_date` by preference, which is the NRS death date unless this is missing, in which case it is the CHI death date, otherwise try the CHI death date.