Amended NSU syntax
Improved comments.
Altered so that any data in the CHI file will be used if it is missing from an existing (non-NSU) record. This won't affect many as most of the service-users have already been filtered from the NSU file.
Had to omit doing the same with `gender` as it is only present in the 19/20 file