Add `it_extract` parameter to IT extract path functions.
The `it_extract` parameter can be used to specify a specific extract for example when using an extract that hasn't been updated recently.
Also expanded the PIS function so that it works independently of a specific extract number. It will now check for a PIS extract for that year with the given reference and return that, regardless of what extract number was used for it.