Replace the SPSS code with some comments and code which will create an error
This doesn't delete the file, so it will still be there as a reminder to use the R code but it can't accidentally be used anymore.

If it is used e.g. in existing production jobs it will create an error and print the messages to the output.

We still need the SPSS test file.

It's also worth noting that all the SPSS code will be kept in the git history.