BUG - fix identification of Care Home CIJ records.
This code highlights the first Care Home episode in the CIJ so that we only count things once. However for people with no Care Home episodes at all (everyone in older years) it was flagging their first episode (in the whole file) as `1` and so they were being counted as having a CH CIJ despite having no CH episodes.