Fix permissions on `write_*` functions
I was getting an error again here, by the same logic we used in `write_xlsx_tests` we should only try and change the permissions if we're the owner i.e. we just create the file.

I think the error message was being confused by the way we pipe the data to `write_zsav` followed by `write_rds`, and how R is lazily executing that pipe (it goes inside out). If you run the `write_*` bits seperately the error message is clearer.

This will need 2 people to test it properly!