Add a (hopefully) helpful warning for non-20C years
I'm not sure if the functionality is wrong per se as technically the year 1917 would be 1917/18 or 1718 in our short form...

That said I don't see a use for it in our work so it will likely be unintended. This warning should highlight any mistake if indeed it is one.