#301 Replace `ooh_cc` with `ooh_case_id`
`ooh_cc` was a bit misleading, both in name and in function. It doesn't help with identifying the same or different cases between years.

`ooh_case_id` replaces it as a unique ID for cases. It does mean we need to do a bit of extra processing in the indiv file to get the unique number of cases.