SC Alarms Telecare Data (#325)
* Use the 'snapshot' versions of the Social Care tables (#203)

* Use snapshot data in care homes

* remove sc_latest_validated_period macro

* Use snapshot data in homecare

* Use snapshot data - Demog/client file

Co-authored-by: Jennifer Thom <jennifer.thom2@phs.scot>

* Update SPD and SIMD macros

New 2022 file

* Update latest/previous update macros

Now Jun_2022 and Mar_2022

* Change `cij_marker` to numeric on read in.

B01 (acute), B02 (maternity) and B04 (mental health). These are the same for all years.

I made the changes in RStudio which has stripped whitespace from the end of the lines on saving.

* Change usage of `cij_marker` in C01

* Change `cij_marker` from string to numeric in C02 (DD linking)

Previously we set an empty string to be 'user missing', now the `cij_marker` is a numeric, any missing values will be 'system missing'.

* Increase the length of `cij_marker` as a string in C07

* Use `cij_marker` as a numeric in D01

* Missed cij_marker string to sysmis in C02 (DD linking)

* BUG - use Alter type (AMIN)

problem with running homelessness linkage in C03 where the type is more than string length 1

* Add June 2022 update to NEWS.md

* Replace the SPSS code with some comments and code which will create an error

This doesn't delete the file, so it will still be there as a reminder to use the R code but it can't accidentally be used anymore.

If it is used e.g. in existing production jobs it will create an error and print the messages to the output.

We still need the SPSS test file.

It's also worth noting that all the SPSS code will be kept in the git history.

* Update NEWS.md

* Fix the CH costs year

The year in the spread sheet (e.g. 2018) refers to the end of the year, therefore the 2018 column contains costs for 2017/18, not 2019/20 as we were using.

* BUG - Fix overcounting of preventable beddays

We were previously counting the beddays for the whole year on every episode. So if a CIJ had 2 episodes it would have 2X the correct number of beddays etc.

* use zsav instead of sav for SPD file

* Update NEWS.md

* Update NEWS.md

* Add home care costs (#217)

* Create lookup for Home Care costs

* Fix bug - should be removing bad dates

* Rename fixed hours variable to make pivot simpler

* Include cost variables

Match from the costs lookup and multiply by hours.

* Use maximum compression for write out

* Style

* Add cost variables to SPSS yearly home care code

* style SPSS

* Add Home Care cost variables to the episode file

* Add Home Care cost variables to the Individual file.

* specify health_board.x in facet wrap
This was preventing the charts from populating on the last graph

* Declare hc cost vars in older years (ep file)

* Declare hc cost vars in older years (indiv file)

* Use latest_update object

Co-authored-by: Jennifer Thom <jennifer.thom@phs.scot>
Co-authored-by: Jennifer Thom <jennifer.thom2@phs.scot>

* Correct erroneous reference to COSLA (#222)

Also, fix some spelling mistakes 🙄

* BUG - change string to sysmis

* Update NEWS.md

* Correctly uplift costs for future years in CH costs

* Update GP Cluster script to latest update

* Fix minor bug - two postcodes were lower case

Converting to upper case still didn't make them match but it doesn't hurt either!

* Change SPD file path to `zsav`

* Create simple tests for the GP Prac lookup

Inspired by the Postcode tests - can be used for looking into the issue of missing Practice names.

* Rename SPD tests script so it sorts after the main script

* Fix for client file variable names

It seems `elderly_frail` has been renamed `elder_frail` in the snapshot as compared to the 'full' version. I suspect that's a mistake so I just renamed the variable on read-in back to `elderly_frail` rather than edit it in the rest of our code.

* Update DD period

* Make changes to GLS coding (#227)

* Make changes to GLS coding

Option 1 - Not sure if this will be implemented but this is the most likely I think.

This will need replicating to the R code if/when we implement it.

* Update NEWS.md

* Take the latest 20 quarters of Cluster data.

Keep only the latest practice name/cluster for each practice.

Can only take 20 because of issues with the open data - this should either be fixed on the platform or with an update to the `phsopendata` package. See [this issue](https://github.com/Public-Health-Scotland/phsopendata/issues/8#issuecomment-1117265704).

* Minor - tidy up whitespace in tests file

* Keep some extra variables in the GP lookup but don't use them in SLF

* Apply changes for GP prac in the episode file too

* Update NEWS.md

* Use missed Snapshot data

A-T and SDS for 2017/18 missed

* Update home care latest_update to June 2022

* Update IT extract Ref

* BUG - Partially revert changes to GP prac matching

I messed it up as I didn't read the code properly!

* Update cij_ppa logic for a numeric cij_marker

* Update NEWS.md

* Update monthly cost variables for all changed costs 

The main issue this change fixes is that previously only the monthly variables for A&A were being updated after cost_total_net was for both A&A and FV. This change fixes that.

It also adds some extra selections to make extremely clear what is being affected.

With the new year selector we could in theory run this code on every year as it would only actually do anything for `year >= "1819"`, this is like the R version.

* Update NEWS.md

* Use the 1415 NSU cohort in the indiv file

We missed this when we added the 1415 cohort!

* BUG - fix identification of Care Home CIJ records.

This code highlights the first Care Home episode in the CIJ so that we only count things once. However for people with no Care Home episodes at all (everyone in older years) it was flagging their first episode (in the whole file) as `1` and so they were being counted as having a CH CIJ despite having no CH episodes.

* Update NEWS.md

* Draft Alarms Telecare script
WIP

* apply styling

* recode sc_send_lca

* keep only R code

* apply dplyr:: binding

* New Alarms Telecare all years data script

* Use snapshot data

* AT queries

* remove validated period code and use snapshot

* use double brackets

* valid chi

* use Rmarkdown code

* R markdown output

* use string instead of numeric

* Update AT scripts
General tidy up/cleaning

* Add sc_latest_submission

* Alarms Telecare tweaks - WIP

* fix bad periods
2020 ~ 2020Q4, 2021 ~ 2021Q4

* remove unused packages

* add lubridate to library

* Merge records across quarters

* Tidy up script

* Add year selection parameters for AT

* Update CH selection
Use NOT(old_open_record)

* Style code

* Deal with multiple sc_id from sending_loc CHI
Taken from CH logic

* Add comments for merging/list vars in summarise

* Remove loaded packages

* use correct compression

* Use the combined SC functions file

* Add the SC dir to the functions script (not as  a function)

* Add `latest_update` to the SC functions

* Style SC functions

* declare sc_dir path

* Add `start_fy` to SC functions

* Replace missing start date with start FY

* Update binding and compression

* Remove data queries script

* Remove latest_sub < FY logic

* Remove age
Not needed here now as we derive this later in the process

* Move SPSS script to 21/22 folder

* Style code

Co-authored-by: Jennifer Thom <jennifer.thom2@phs.scot>
Co-authored-by: James McMahon <james.mcmahon@phs.scot>