302 updates and changes to gp out of hours (#358)
* #300 Add `ConsultationTypeUnmapped` and map the COVID values

This is based on coversations with the Unscheduled care team and is in-line with how they do this in their publication.

* #300 Exclude some consultations

From discussions with the Unscheduled Care team, these consultations shouldn't be included as they are wrongly submitted Flow navigation centre data.

* #301 Replace `ooh_cc` with `ooh_case_id`

`ooh_cc` was a bit misleading, both in name and in function. It doesn't help with identifying the same or different cases between years.

`ooh_case_id` replaces it as a unique ID for cases. It does mean we need to do a bit of extra processing in the indiv file to get the unique number of cases.

Co-authored-by: Jennit07 <67372904+Jennit07@users.noreply.github.com>