Changes due to `tidyselect` 1.2.0
In the latest version of `tidyselect` 1.2.0 they deprecated the use of `.data$` in 'selection contexts' this means things like `dplyr::select` and `across`, and probably some others too!

This was causing a lot of warnings to appear for our tests.

I've gone through and replaced all the instances I could find. The tests are now (still) passing with no warnings.

I'm aware that not all our code is covered by the tests though so please check to see if I've missed or broken anything.