Quick updates to make the pkgdown site work (#568)
* Quick updates to make the pkgdown site work

Quite a few of the more recent functions were missing from the index, which was causing the site to not build. I've very roughly / quickly added them all so that now the site builds. It's definitely not the best way to present the documentation so will need revisiting at some point!

* Style code

---------

Co-authored-by: Moohan <Moohan@users.noreply.github.com>